# perf: add caches to commit_has_blob and commit_has_tree