Closes #191 

### Commit has tree

Before:

```
BenchmarkCommitHasTree/commit_has_tree-4                    3000            519234 ns/op          297322 B/op       1276 allocs/op
PASS
ok      github.com/src-d/gitquery/internal/function     2.154s
```

After:

```
BenchmarkCommitHasTree/commit_has_tree-4                 1000000              1036 ns/op             240 B/op          7 allocs/op
PASS
ok      github.com/src-d/gitquery/internal/function     1.593s
```

### Commit has blob

Before:

```
BenchmarkCommitHasBlob/commit_has_blob-4                     500           2605190 ns/op         1143268 B/op       1768 allocs/op
PASS
ok      github.com/src-d/gitquery/internal/function     2.105s
```

After:

```
BenchmarkCommitHasBlob/commit_has_blob-4                 1000000              1013 ns/op             240 B/op          7 allocs/op
PASS
ok      github.com/src-d/gitquery/internal/function     1.559s
```

**DISCLAIMER:** benchmark here is biased because the same commits keep being asked again and again. In a regular production query the improvement would likely be much lower than it is here. Still an improvement, nonetheless.