@erizocosmico 2018-03-26
To no add it twice, the commit adding golang-lru to vendor is in #198