@erizocosmico 2018-04-03
Changed from ARCCache to TwoQueueCache